Fix __iter__ and .shape
